Start adding context. Not working. Work in progress
