+++ src/Components/index.js
@@ -1,4 +1,5 @@
 export {default as ProjectCard } from './ProjectCard/ProjectCard.component';
+export {default as ProjectCardHolder } from './ProjectCard/ProjectCardHolder.component'
 export {default as SignIn } from './SignIn/SignIn.component';
 export {default as SignUp } from './SignUp/SignUp.component';
 export {default as ProjectForm } from './ProjectForm/ProjectForm.component'
