--- src/fakeData.js
@@ -1,23 +1,0 @@
-export const fakeData = [{
-    company: "Apolo FX",
-    projectTitle: "Arduino Developer for Guitar Pedal sound effect development",
-    projectDescription: "Apolo FX develops professional pedal effects for electric guitars. We are currently developing a new product that requires programing some effects with Arduino in C++",
-    roles: "Arduino Developer, C++ Developer",
-    skills: "Arduino, C++, C, Electronics",
-    duration: "> 6 months",
-    postDate: "21/12/2021",
-    expirationDate: "1/2/2022",
-    id: 123
-},
-{
-    company: "Nowitis",
-    projectTitle: "Development of new Javascript features in Marketing webpage",
-    projectDescription: "We are fast growing marketing agency. We have developed our website with Wix and now we would like to add some features and animations. If the collaborator is willing to, we would appreciate a fully rebuild of the website on whichever language the developer offers",
-    roles: "Web developer",
-    skills: "Web development, Javascript, React.js, Angular, Vue.js, PHP",
-    duration: "3 months",
-    postDate: "23/11/2021",
-    expirationDate: "21/3/2022",
-    id: 124
-}
-]
